Add a Fixed Title Block at the top Closes: #1189
Add a fixed title block which cannot be deleted or removed by the end user.
This functionality can be enabled by setting the 'fixedTitleBlock' property to true.
It can also be intiialized with an Object containing the placeholder property which
will be used as the placeholder text for the title block.
The block uses a textarea with autosize module to grow as the text size increase.
The 'OutputData' class changes and adds another property called 'title'
which contains the title provided by the user.
It cannot be added to the 'blocks' property itself because then it would have not been
at a fixed position and removable by the end user.
There is not way found to sanitize the data as EditorJS.Sanitizer was not a property and
in order to intialize Sanitizer module Editor config object requires which is only present in
core.ts and codex.ts and cannot be sent to saver making it not possible to save the data.
The documentation will be added once the change is approved.